refactor: rental business folders
