+++ packages/web/src/features/my/frames/MyServiceFrame.tsx
@@ -10,11 +10,11 @@ import MyActivityCertificateTable from "@sparcs-clubs/web/features/my/components
 import MyCommonSpaceTable from "@sparcs-clubs/web/features/my/components/MyCommonSpaceTable";
 import MyPrintingTable from "@sparcs-clubs/web/features/my/components/MyPrintingTable";
 import MyRentalTable from "@sparcs-clubs/web/features/my/components/MyRentalTable";
+import { useGetMyRentals } from "@sparcs-clubs/web/features/rental-business/service/getMyRentals";
 
 import { useGetMyActivityCertificate } from "../services/getMyActivityCertificate";
 import { useGetMyCommonSpace } from "../services/getMyCommonSpace";
 import { useGetMyPrinting } from "../services/getMyPrinting";
-import { useGetMyRentals } from "../services/getMyRentals";
 
 const MyServiceFrame: React.FC = () => {
   // TODO: 실제 필요한 값으로 바꾸기
