+++ packages/web/src/features/rental-business/frames/RentalDetailFrame.tsx
@@ -5,7 +5,7 @@ import {
 } from "@sparcs-clubs/web/common/components/ListItem";
 import Typography from "@sparcs-clubs/web/common/components/Typography";
 
-const ManageRentalDetailFrame = () => (
+const RentalDetailFrame = () => (
   //   TODO: 아래 정보들 백 연결하기
   <FlexWrapper direction="column" gap={20}>
     <FlexWrapper direction="column" gap={16}>
@@ -50,4 +50,4 @@ const ManageRentalDetailFrame = () => (
     </FlexWrapper>
   </FlexWrapper>
 );
-export default ManageRentalDetailFrame;
+export default RentalDetailFrame;
